[fronius] Move FroniusHttpUtil & FroniusCommunicationException to API package
